refactor(google-workspace): unify service layout and hardcode tool annotations
Every service now follows the same flat shape (index.ts + tools.ts, plus
optional schemas.ts / helpers/), and every ToolDefinition carries its
annotations literal next to the name/description/inputSchema instead of
deriving them from a name-prefix classifier at startup.

- Move src/tools/{drive,docs,sheets,slides}.ts -> src/services/<svc>/tools.ts;
  index.ts becomes an 11-line ServiceModule wrapper for each.
- Flatten gmail: split index.ts into tools.ts (registry + dispatcher) and a
  thin index.ts wrapper.
- Flatten calendar: collapse 9 BaseToolHandler subclasses into runX functions
  inside tools.ts, consolidate schemas/{inputs,types}.ts into schemas.ts, and
  pull shared logic out into helpers/{datetime,format,api,batch,recurring}.ts.
- Drop src/annotations.ts and src/tools/. Annotations are now hardcoded with
  a minimal shape: { readOnlyHint: true } for read-only tools,
  { destructiveHint: false } for additive tools (overrides MCP SDK's true
  default), { destructiveHint: true } for destructive tools.
- Update README to document the new layout and annotation convention.

Verified: tsc --noEmit + npm run build both pass; the [R]/[]/[D] badge
output of scripts/list-tools.mjs is unchanged across all 123 tools.

Made-with: Cursor
