+++ google-workspace/src/services/calendar/helpers/batch.ts
@@ -1,4 +1,9 @@
-import { OAuth2Client } from "google-auth-library";
+/**
+ * Google Calendar batch HTTP request helper. Carried over from the upstream
+ * google-calendar-mcp BatchRequestHandler with light cleanups; behavior is
+ * unchanged.
+ */
+import { OAuth2Client } from 'google-auth-library';
 
 export interface BatchRequest {
   method: string;
@@ -25,7 +30,7 @@ export class BatchRequestError extends Error {
   constructor(
     message: string,
     public errors: BatchError[],
-    public partial: boolean = false
+    public partial: boolean = false,
   ) {
     super(message);
     this.name = 'BatchRequestError';
@@ -33,49 +38,51 @@ export class BatchRequestError extends Error {
 }
 
 export class BatchRequestHandler {
-  private readonly batchEndpoint = "https://www.googleapis.com/batch/calendar/v3";
+  private readonly batchEndpoint = 'https://www.googleapis.com/batch/calendar/v3';
   private readonly boundary: string;
   private readonly maxRetries = 3;
-  private readonly baseDelay = 1000; // 1 second
+  private readonly baseDelay = 1000;
 
   constructor(private auth: OAuth2Client) {
-    this.boundary = "batch_boundary_" + Date.now();
+    this.boundary = 'batch_boundary_' + Date.now();
   }
 
   async executeBatch(requests: BatchRequest[]): Promise<BatchResponse[]> {
-    if (requests.length === 0) {
-      return [];
-    }
-
+    if (requests.length === 0) return [];
     if (requests.length > 50) {
       throw new Error('Batch requests cannot exceed 50 requests per batch');
     }
-
     return this.executeBatchWithRetry(requests, 0);
   }
 
-  private async executeBatchWithRetry(requests: BatchRequest[], attempt: number): Promise<BatchResponse[]> {
+  private async executeBatchWithRetry(
+    requests: BatchRequest[],
+    attempt: number,
+  ): Promise<BatchResponse[]> {
     try {
       const batchBody = this.createBatchBody(requests);
       const token = await this.auth.getAccessToken();
-      
+
       const response = await fetch(this.batchEndpoint, {
-        method: "POST",
+        method: 'POST',
         headers: {
-          "Authorization": `Bearer ${token.token}`,
-          "Content-Type": `multipart/mixed; boundary=${this.boundary}`
+          Authorization: `Bearer ${token.token}`,
+          'Content-Type': `multipart/mixed; boundary=${this.boundary}`,
         },
-        body: batchBody
+        body: batchBody,
       });
 
       const responseText = await response.text();
 
-      // Handle rate limiting with retry
       if (response.status === 429 && attempt < this.maxRetries) {
         const retryAfter = response.headers.get('Retry-After');
-        const delay = retryAfter ? parseInt(retryAfter) * 1000 : this.baseDelay * Math.pow(2, attempt);
-        
-        process.stderr.write(`Rate limited, retrying after ${delay}ms (attempt ${attempt + 1}/${this.maxRetries})\n`);
+        const delay = retryAfter
+          ? parseInt(retryAfter) * 1000
+          : this.baseDelay * Math.pow(2, attempt);
+
+        process.stderr.write(
+          `Rate limited, retrying after ${delay}ms (attempt ${attempt + 1}/${this.maxRetries})\n`,
+        );
         await this.sleep(delay);
         return this.executeBatchWithRetry(requests, attempt + 1);
       }
@@ -83,36 +90,42 @@ export class BatchRequestHandler {
       if (!response.ok) {
         throw new BatchRequestError(
           `Batch request failed: ${response.status} ${response.statusText}`,
-          [{
-            statusCode: response.status,
-            message: `HTTP ${response.status}: ${response.statusText}`,
-            details: responseText
-          }]
+          [
+            {
+              statusCode: response.status,
+              message: `HTTP ${response.status}: ${response.statusText}`,
+              details: responseText,
+            },
+          ],
         );
       }
 
       return this.parseBatchResponse(responseText);
     } catch (error) {
-      if (error instanceof BatchRequestError) {
-        throw error;
-      }
-      
-      // Retry on network errors
+      if (error instanceof BatchRequestError) throw error;
+
       if (attempt < this.maxRetries && this.isRetryableError(error)) {
         const delay = this.baseDelay * Math.pow(2, attempt);
-        process.stderr.write(`Network error, retrying after ${delay}ms (attempt ${attempt + 1}/${this.maxRetries}): ${error instanceof Error ? error.message : 'Unknown error'}\n`);
+        process.stderr.write(
+          `Network error, retrying after ${delay}ms (attempt ${attempt + 1}/${
+            this.maxRetries
+          }): ${error instanceof Error ? error.message : 'Unknown error'}\n`,
+        );
         await this.sleep(delay);
         return this.executeBatchWithRetry(requests, attempt + 1);
       }
-      
-      // Handle network or auth errors
+
       throw new BatchRequestError(
-        `Failed to execute batch request: ${error instanceof Error ? error.message : 'Unknown error'}`,
-        [{
-          statusCode: 0,
-          message: error instanceof Error ? error.message : 'Unknown error',
-          details: error
-        }]
+        `Failed to execute batch request: ${
+          error instanceof Error ? error.message : 'Unknown error'
+        }`,
+        [
+          {
+            statusCode: 0,
+            message: error instanceof Error ? error.message : 'Unknown error',
+            details: error,
+          },
+        ],
       );
     }
   }
@@ -120,51 +133,54 @@ export class BatchRequestHandler {
   private isRetryableError(error: any): boolean {
     if (error instanceof Error) {
       const message = error.message.toLowerCase();
-      return message.includes('network') || 
-             message.includes('timeout') || 
-             message.includes('econnreset') ||
-             message.includes('enotfound');
+      return (
+        message.includes('network') ||
+        message.includes('timeout') ||
+        message.includes('econnreset') ||
+        message.includes('enotfound')
+      );
     }
     return false;
   }
 
   private sleep(ms: number): Promise<void> {
-    return new Promise(resolve => setTimeout(resolve, ms));
+    return new Promise((resolve) => setTimeout(resolve, ms));
   }
 
   private createBatchBody(requests: BatchRequest[]): string {
-    return requests.map((req, index) => {
-      const parts = [
-        `--${this.boundary}`,
-        `Content-Type: application/http`,
-        `Content-ID: <item${index + 1}>`,
-        "",
-        `${req.method} ${req.path} HTTP/1.1`
-      ];
+    return (
+      requests
+        .map((req, index) => {
+          const parts = [
+            `--${this.boundary}`,
+            `Content-Type: application/http`,
+            `Content-ID: <item${index + 1}>`,
+            '',
+            `${req.method} ${req.path} HTTP/1.1`,
+          ];
 
-      if (req.headers) {
-        Object.entries(req.headers).forEach(([key, value]) => {
-          parts.push(`${key}: ${value}`);
-        });
-      }
+          if (req.headers) {
+            Object.entries(req.headers).forEach(([key, value]) => {
+              parts.push(`${key}: ${value}`);
+            });
+          }
 
-      if (req.body) {
-        parts.push("Content-Type: application/json");
-        parts.push("");
-        parts.push(JSON.stringify(req.body));
-      }
+          if (req.body) {
+            parts.push('Content-Type: application/json');
+            parts.push('');
+            parts.push(JSON.stringify(req.body));
+          }
 
-      return parts.join("\r\n");
-    }).join("\r\n\r\n") + `\r\n--${this.boundary}--`;
+          return parts.join('\r\n');
+        })
+        .join('\r\n\r\n') + `\r\n--${this.boundary}--`
+    );
   }
 
   private parseBatchResponse(responseText: string): BatchResponse[] {
-    // First, try to find boundary from Content-Type header in the response
-    // Google's responses typically have boundary in the first few lines
     const lines = responseText.split(/\r?\n/);
-    let boundary = null;
-    
-    // Look for Content-Type header with boundary in the first few lines
+    let boundary: string | null = null;
+
     for (let i = 0; i < Math.min(10, lines.length); i++) {
       const line = lines[i];
       if (line.toLowerCase().includes('content-type:') && line.includes('boundary=')) {
@@ -175,45 +191,33 @@ export class BatchRequestHandler {
         }
       }
     }
-    
-    // If not found in headers, try to find boundary markers in the content
+
     if (!boundary) {
       const boundaryMatch = responseText.match(/--([a-zA-Z0-9_-]+)/);
-      if (boundaryMatch) {
-        boundary = boundaryMatch[1];
-      }
+      if (boundaryMatch) boundary = boundaryMatch[1];
     }
-    
+
     if (!boundary) {
       throw new Error('Could not find boundary in batch response');
     }
-    
-    // Split by boundary markers
+
     const parts = responseText.split(`--${boundary}`);
-    
     const responses: BatchResponse[] = [];
-    
-    // Skip the first part (before the first boundary) and the last part (after final boundary with --)
+
     for (let i = 1; i < parts.length; i++) {
       const part = parts[i];
-      
-      // Skip empty parts or the final boundary marker
       if (part.trim() === '' || part.trim() === '--' || part.trim().startsWith('--')) continue;
-      
+
       const response = this.parseResponsePart(part);
-      if (response) {
-        responses.push(response);
-      }
+      if (response) responses.push(response);
     }
-    
+
     return responses;
   }
 
   private parseResponsePart(part: string): BatchResponse | null {
-    // Handle both \r\n and \n line endings
     const lines = part.split(/\r?\n/);
-    
-    // Find the HTTP response line (look for "HTTP/1.1")
+
     let httpLineIndex = -1;
     for (let i = 0; i < lines.length; i++) {
       if (lines[i].startsWith('HTTP/1.1')) {
@@ -221,27 +225,22 @@ export class BatchRequestHandler {
         break;
       }
     }
-
     if (httpLineIndex === -1) return null;
 
-    // Parse status code from HTTP response line
     const httpLine = lines[httpLineIndex];
     const statusMatch = httpLine.match(/HTTP\/1\.1 (\d+)/);
     if (!statusMatch) return null;
-    
-    const statusCode = parseInt(statusMatch[1]);
 
-    // Parse response headers (start after HTTP line, stop at empty line)
+    const statusCode = parseInt(statusMatch[1]);
     const headers: Record<string, string> = {};
     let bodyStartIndex = httpLineIndex + 1;
-    
+
     for (let i = httpLineIndex + 1; i < lines.length; i++) {
       const line = lines[i];
       if (line.trim() === '') {
         bodyStartIndex = i + 1;
         break;
       }
-      
       const colonIndex = line.indexOf(':');
       if (colonIndex > 0) {
         const key = line.substring(0, colonIndex).trim();
@@ -250,37 +249,25 @@ export class BatchRequestHandler {
       }
     }
 
-    // Parse body - everything after the empty line following headers
     let body: any = null;
     if (bodyStartIndex < lines.length) {
-      // Collect all body lines, filtering out empty lines at the end
-      const bodyLines = [];
-      for (let i = bodyStartIndex; i < lines.length; i++) {
-        bodyLines.push(lines[i]);
-      }
-      
-      // Remove trailing empty lines
+      const bodyLines: string[] = [];
+      for (let i = bodyStartIndex; i < lines.length; i++) bodyLines.push(lines[i]);
       while (bodyLines.length > 0 && bodyLines[bodyLines.length - 1].trim() === '') {
         bodyLines.pop();
       }
-      
       if (bodyLines.length > 0) {
         const bodyText = bodyLines.join('\n');
         if (bodyText.trim()) {
           try {
             body = JSON.parse(bodyText);
           } catch {
-            // If JSON parsing fails, return the raw text
             body = bodyText;
           }
         }
       }
     }
 
-    return {
-      statusCode,
-      headers,
-      body
-    };
+    return { statusCode, headers, body };
   }
-} 
+}
