+++ google-workspace/src/services/docs/index.ts
@@ -1,16 +1,5 @@
-import type { ServiceModule, ToolContext, ToolResult } from '../../types.js';
-import { annotateAll } from '../../annotations.js';
-import * as docsTools from '../../tools/docs.js';
-
-const toolDefinitions = annotateAll(docsTools.toolDefinitions);
-
-async function handleTool(
-  name: string,
-  args: Record<string, any>,
-  ctx: ToolContext,
-): Promise<ToolResult | null> {
-  return docsTools.handleTool(name, args, ctx);
-}
+import type { ServiceModule } from '../../types.js';
+import { toolDefinitions, handleTool } from './tools.js';
 
 const docsService: ServiceModule = {
   key: 'docs',
