+++ google-workspace/src/services/drive/index.ts
@@ -1,16 +1,5 @@
-import type { ServiceModule, ToolContext, ToolResult } from '../../types.js';
-import { annotateAll } from '../../annotations.js';
-import * as driveTools from '../../tools/drive.js';
-
-const toolDefinitions = annotateAll(driveTools.toolDefinitions);
-
-async function handleTool(
-  name: string,
-  args: Record<string, any>,
-  ctx: ToolContext,
-): Promise<ToolResult | null> {
-  return driveTools.handleTool(name, args, ctx);
-}
+import type { ServiceModule } from '../../types.js';
+import { toolDefinitions, handleTool } from './tools.js';
 
 const driveService: ServiceModule = {
   key: 'drive',
