+++ google-workspace/src/services/sheets/index.ts
@@ -1,16 +1,5 @@
-import type { ServiceModule, ToolContext, ToolResult } from '../../types.js';
-import { annotateAll } from '../../annotations.js';
-import * as sheetsTools from '../../tools/sheets.js';
-
-const toolDefinitions = annotateAll(sheetsTools.toolDefinitions);
-
-async function handleTool(
-  name: string,
-  args: Record<string, any>,
-  ctx: ToolContext,
-): Promise<ToolResult | null> {
-  return sheetsTools.handleTool(name, args, ctx);
-}
+import type { ServiceModule } from '../../types.js';
+import { toolDefinitions, handleTool } from './tools.js';
 
 const sheetsService: ServiceModule = {
   key: 'sheets',
