+++ google-workspace/src/services/slides/index.ts
@@ -1,16 +1,5 @@
-import type { ServiceModule, ToolContext, ToolResult } from '../../types.js';
-import { annotateAll } from '../../annotations.js';
-import * as slidesTools from '../../tools/slides.js';
-
-const toolDefinitions = annotateAll(slidesTools.toolDefinitions);
-
-async function handleTool(
-  name: string,
-  args: Record<string, any>,
-  ctx: ToolContext,
-): Promise<ToolResult | null> {
-  return slidesTools.handleTool(name, args, ctx);
-}
+import type { ServiceModule } from '../../types.js';
+import { toolDefinitions, handleTool } from './tools.js';
 
 const slidesService: ServiceModule = {
   key: 'slides',
